fix psalm warning

+++ src/Binary.php
@@ -54,7 +54,7 @@ class Binary
     public static function safeStrlen(string $str): int
     {
         if (\function_exists('mb_strlen')) {
-            return (int) mb_strlen($str, '8bit');
+            return mb_strlen($str, '8bit');
         }
 
         return \strlen($str);
